+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2020-11-07
+- Added `rustpat` Rust wrapper
+- **[BREAKING CHANGE]** Replaced `pat_open` with `pat_init` in `libpat`
+- **[BREAKING CHANGE]** Removed `PAT` struct from `libpat`
+
 ## [0.4.1] - 2020-10-28
 ### Changed
 - Updated dependencies